src/requirements.txt
python-dotenv
boto3
xhtml2pdf
requests
pandas